Switch data rows container to ng-show
Fixes #131. The problem was that the scroll handler was only being
assigned while the ng-if condition was false, so the selector couldn't
find the stuff it was supposed to be selecting. I think ng-show is
appropriate, given that its conditions should almost always be true
anyway - we aren't really saving much on performance by making it ng-if.